# tests/fixtures/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskBatchSubmit": {
        "description": "Request body for POST /tasks/batch.",
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskBatchItem"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "tasks"
        ],
        "title": "TaskBatchSubmit",
        "type": "object"
      },
      "TaskPriorityUpdate": {
        "description": "Request body for PATCH /tasks/{task_id}.",
        "properties": {
          "priority": {
            "title": "Priority",
            "type": "integer"
          }
        },
        "required": [
          "priority"
        ],
        "title": "TaskPriorityUpdate",
        "type": "object"
      },
      "TaskSubmit": {
        "properties": {
          "depends_on": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Depends On",
            "type": "array"
          },
          "inherit_priority": {
            "default": true,
            "title": "Inherit Priority",
            "type": "boolean"
          },
          "max_retries": {
            "default": 0,
            "title": "Max Retries",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "priority": {
            "default": 0,
            "title": "Priority",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "reply_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply To"
          },
          "required_tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Required Tags",
            "type": "array"
          },
          "target_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Agent"
          },
          "target_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Group"
          },
          "ttl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "TaskSubmit",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tasks": {
      "get": {
        "description": "Return all tasks: currently queued, in-progress, and completed/failed.\n\nCombines the pending queue, currently dispatched (in-progress) tasks,\nand per-agent history into a single flat list.  Use ``skip`` and\n``limit`` query params for pagination.\n\nDesign reference:\n- AWS SQS message visibility / dead-letter queue listing\n- DESIGN.md \u00a710.21 (v0.26.0)",
        "operationId": "list_tasks_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Tasks Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List all tasks (active + completed)"
      },
      "post": {
        "operationId": "submit_task_limited_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Task Limited Tasks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a new task"
      }
    },
    "/tasks/batch": {
      "post": {
        "description": "Submit a list of tasks atomically.\n\nAll tasks in the batch are validated before any are enqueued.  If the\nrequest body is malformed, FastAPI returns 422 before this handler runs.\n\nDesign reference:\n- adidas API Guidelines \"Batch Operations\"\n  https://adidas.gitbook.io/api-guidelines/rest-api-guidelines/execution/batch-operations\n- PayPal Batch API (Medium, PayPal Tech Blog)\n  https://medium.com/paypal-tech/batch-an-api-to-bundle-multiple-paypal-rest-operations-6af6006e002",
        "operationId": "submit_tasks_batch_tasks_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskBatchSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Tasks Batch Tasks Batch Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit multiple tasks in one request"
      }
    },
    "/tasks/{task_id}": {
      "delete": {
        "description": "Cancel *task_id* whether it is queued or currently in-progress.\n\nDesign references:\n- Kubernetes ``kubectl delete pod`` \u2014 REST DELETE on a resource URI\n- POSIX SIGTERM/SIGKILL model; Go context.Context cancellation\n- DESIGN.md \u00a710.22 (v0.27.0)",
        "operationId": "delete_task_tasks__task_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Task Tasks  Task Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel a task by ID (queued or in-progress)"
      },
      "get": {
        "description": "Return the status and details of a specific task by its ID.\n\nDesign reference: DESIGN.md \u00a710.21 (v0.26.0); DESIGN.md \u00a710.24 (v0.29.0)",
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Task Tasks  Task Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a specific task by ID"
      },
      "patch": {
        "description": "Update the priority of a task that is still in the pending queue.",
        "operationId": "update_task_priority_tasks__task_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskPriorityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Task Priority Tasks  Task Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a pending task's priority"
      }
    },
    "/tasks/{task_id}/cancel": {
      "post": {
        "description": "Remove *task_id* from the pending queue and discard it.\n\nReturns:\n- ``{\"cancelled\": true, \"task_id\": ..., \"status\": \"cancelled\"}``\n  if the task was successfully removed from the queue.\n- ``{\"cancelled\": false, \"task_id\": ..., \"status\": \"already_dispatched\"}``\n  if the task was not in the pending queue (already dispatched or\n  currently in-flight).\n- ``404`` if the task ID has never been submitted or tracked.\n\nDesign reference: DESIGN.md \u00a711 (v0.17.0) \u2014 task cancellation.",
        "operationId": "cancel_task_tasks__task_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Task Tasks  Task Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel a pending task"
      }
    },
...
  }
}